refactor program to raider.py
Begin support for reading ARIA files
